renamed enum-labels connector to property-panel-ux as it now shows off all possible UI features
