--- enum-labels/connector.json
@@ -1,14 +1,0 @@
-{
-    "name": "enum-labels",
-    "version": "1.0",
-    "title": "Enum Labels",
-    "description": "An example connector of how to use enum labels",
-    "service": {
-        "name": "json-placeholder",
-        "version": "1"
-    },
-    "tags": [
-        "service"
-    ],
-    "isTrigger": false
-}
