--- enum-labels/src/create_post/input.json
@@ -1,1 +1,0 @@
-{"type":"object","properties":{"title":{"enumLabels":["Reminder","To do","Journal","Task"],"type":"string","enum":["reminder","todo","journal","task"]},"body":{"type":"string"},"userId":{"type":"number"}},"additionalProperties":false,"required":["body","title","userId"],"$schema":"http://json-schema.org/draft-07/schema#","advanced":[]}
