--- enum-labels/src/create_post/input.ts
@@ -1,15 +1,0 @@
-export type CreatePostInput = {
-	title: PostType;
-	body: string;
-	userId: number;
-};
-
-/**
- * @enumLabels Reminder, To do, Journal, Task
- */
-export enum PostType {
-	reminder = 'reminder',
-	todo = 'todo',
-	journal = 'journal',
-	task = 'task',
-}
