+++ property-panel-ux/connector.json
@@ -1,0 +1,14 @@
+{
+    "name": "property-panel-ux",
+    "version": "1.0",
+    "title": "Property Panel UX",
+    "description": "An example connector showing how to use UI features in the property panel, including: field titles, descriptions (aka tooltips), default values, advanced properties and oneOf (aka dropdowns) field types.",
+    "service": {
+        "name": "json-placeholder",
+        "version": "1"
+    },
+    "tags": [
+        "service"
+    ],
+    "isTrigger": false
+}
