+++ property-panel-ux/src/PropertyPanelUxAuth.ts
@@ -1,0 +1,1 @@
+export type PropertyPanelUxAuth = never; // No Authentication
