+++ property-panel-ux/src/create_post/input.json
@@ -1,0 +1,1 @@
+{"type":"object","properties":{"title":{"enumLabels":["Reminder","To do","Journal","Task"],"title":"Post Type","description":"The type of post to create","default":"task","type":"string","enum":["reminder","todo","journal","task"]},"body":{"title":"Body","description":"The body of the post","default":"Post content goes here","type":"string"},"userId":{"title":"User ID","description":"The ID of the user to create the post for","type":"number"},"owner":{"title":"Owner","description":"The owner of the post (optional, either a non profit or business)","oneOf":[{"title":"Non Profit","type":"object","properties":{"name":{"title":"Non Profit Name","description":"The non profit to create the post for","type":"string"}},"additionalProperties":false,"required":["name"],"advanced":[]},{"title":"Business","type":"object","properties":{"name":{"title":"Business Name","description":"The business to create the post for","type":"string"},"address":{"title":"Business Address","type":"string"}},"additionalProperties":false,"required":["address","name"],"advanced":[]}]}},"additionalProperties":false,"required":["body","title","userId"],"$schema":"http://json-schema.org/draft-07/schema#","advanced":["owner"]}
