+++ property-panel-ux/src/create_post/input.ts
@@ -1,0 +1,61 @@
+export type CreatePostInput = {
+	/**
+	 * @title Post Type
+	 * @description The type of post to create
+	 * @default task
+	 */
+	title: PostType;
+	/**
+	 * @title Body
+	 * @description The body of the post
+	 * @default Post content goes here
+	 */
+	body: string;
+	/**
+	 * @title User ID
+	 * @description The ID of the user to create the post for
+	 */
+	userId: number;
+	/**
+	 * @title Owner
+	 * @description The owner of the post (optional, either a non profit or business)
+	 * @advanced true
+	 */
+	owner?: NonProfit | Business
+};
+
+/**
+ * @title Non Profit
+ */
+export type NonProfit = {
+	/**
+	 * @title Non Profit Name
+	 * @description The non profit to create the post for
+	 */
+	name: string;
+};
+
+/**
+ * @title Business
+ */
+export type Business = {
+	/**
+	 * @title Business Name
+	 * @description The business to create the post for
+	 */
+	name: string;
+	/**
+	 * @title Business Address
+	 */
+	address: string;
+};
+
+/**
+ * @enumLabels Reminder, To do, Journal, Task
+ */
+export enum PostType {
+	reminder = 'reminder',
+	todo = 'todo',
+	journal = 'journal',
+	task = 'task',
+}
